// openapi/arena-sak-innsyn-typer.ts
export interface paths {
    "/arena/innsyn/person": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** Hent personId for en person */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody: {
                content: {
                    "application/json": components["schemas"]["IdentForesporsel"];
                };
            };
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["PersonIdResponse"];
                    };
                };
                /** @description Feil ved henting av personId */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/problem+json": components["schemas"]["HttpProblem"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/arena/innsyn/person/{personId}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Hent personinformasjon for en person */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    personId: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ArenaPersonResponse"];
                    };
                };
                /** @description Feil ved henting av personinformasjon */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/problem+json": components["schemas"]["HttpProblem"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/arena/innsyn/sak/person/{personId}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Hent alle saker for en person */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    personId: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ArenaSakResponse"][];
                    };
                };
                /** @description Feil ved henting av saker */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/problem+json": components["schemas"]["HttpProblem"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/arena/innsyn/sak/{sakid}/detaljert": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Hent detaljert informasjon om en sak på sakId */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    sakid: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ArenaSakDetaljerResponse"];
                    };
                };
                /** @description Sak ikke funnet */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/problem+json": components["schemas"]["HttpProblem"];
                    };
                };
                /** @description Feil ved henting av sak */
                default: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/problem+json": components["schemas"]["HttpProblem"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        ArenaSakResponse: {
            sakId?: string;
            opprettetAar?: string;
            lopenr?: number;
            statuskode?: string;
            statusnavn?: string;
            /** Format: date-time */
            registrertDato?: string;
            /** Format: date-time */
            avsluttetDato?: string;
        };
        ArenaSakDetaljerResponse: {
            sakId?: string;
            opprettetAar: string;
            lopenr: number;
            person: components["schemas"]["ArenaPersonResponse"];
            statuskode: string;
            statusnavn: string;
            /** Format: date-time */
            registrertDato: string;
            /** Format: date-time */
            avsluttetDato?: string;
            /** @default [] */
            vedtak: components["schemas"]["ArenaVedtakMedDetaljerResponse"][];
            telleverkForPerson?: components["schemas"]["TelleverkForPersonResponse"];
            /** @default [] */
            kvoteHistorikk: components["schemas"]["KvotebrukHendelseResponse"][];
            /** Format: date */
            sisteUtbetalingDato?: string;
        };
        ArenaPersonResponse: {
            personId: number;
            fodselsnummer: string;
            fornavn: string;
            etternavn: string;
        };
        ArenaVedtakMedDetaljerResponse: {
            vedtakId?: number;
            lopenrvedtak: number;
            statusKode?: string;
            statusNavn?: string;
            vedtaktypeKode?: string;
            vedtaktypeNavn?: string;
            aktivitetsfaseKode?: string;
            aktivitetsfaseNavn?: string;
            /** Format: date */
            fraOgMed?: string;
            /** Format: date */
            tilDato?: string;
            rettighetkode?: string;
            rettighetnavn?: string;
            utfallkode?: string;
            begrunnelse?: string;
            saksbehandler?: string;
            beslutter?: string;
            relatertVedtak?: number;
            /** @default [] */
            fakta: components["schemas"]["ArenaVedtakfaktaResponse"][];
            /** @default [] */
            "vilk\u00E5rsvurderinger": components["schemas"]["ArenaVilkarsvurderingResponse"][];
        };
        TelleverkForPersonResponse: {
            dagpengePeriodeTeller: number;
            maxPeriodePermittertTellerFisk?: number;
            maxPeriodePermittertTeller?: number;
        };
        KvotebrukHendelseResponse: {
            id: number;
            kvoteTypeKode: string;
            endringsGrunnlag: string;
            antallBevegelse: number;
            posteringTypeKode: string;
            /** Format: date */
            datoHendelse: string;
            resterende: number;
            modUser?: string;
            begrunnelse?: string;
        };
        ArenaVedtakfaktaResponse: {
            kode: string;
            navn: string;
            verdi?: string;
            /** Format: date */
            registrertDato: string;
        };
        ArenaVilkarsvurderingResponse: {
            /** Format: int64 */
            "vilk\u00E5rsvurderingId": number;
            "vilk\u00E5rkode": string;
            begrunnelse?: string;
            vurdertAv?: string;
            "vilk\u00E5rnavn": string;
            erObligatorisk: boolean;
            hjelpetekstUrl?: string;
            lovtekstUrl?: string;
            rundskrivUrl?: string;
            statuskode: string;
            statusnavn: string;
        };
        /** @description Unik identifikator for en person i Arena. */
        PersonIdResponse: {
            id: number;
        };
        IdentForesporsel: {
            /** @description Fødselsnummer eller D-nummer til en person. Må være 11 siffer. */
            ident: string;
        };
        HttpProblem: {
            /**
             * Format: uri
             * @description A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when dereferenced, it provide human-readable documentation for the problem type
             */
            type: string;
            /** @description A short, human-readable summary of the problem type. It SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization */
            title: string;
            /** @description The HTTP status code ([RFC7231], Section 6) generated by the origin server for this occurrence of the problem */
            status: number;
            /** @description A human-readable explanation specific to this occurrence of the problem */
            detail?: string;
            /**
             * Format: uri
             * @description A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced
             */
            instance?: string;
        } & {
            [key: string]: unknown;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}